Updating Build for Frontend
Decided to go with the LoadBalancer approach (why not?) and have refactored the configuration to make sure it reach the reverse proxy.

Spotted a potential issue whereby building the image locally on an M1 Mac can cause errors so I am updating Travis to build the image with a V2 tag.

Link to StackOverflow with the possible cause: https://stackoverflow.com/a/71102144